+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.0.20 (2022-05-06)
 
 - Added key mapping for previous and next shots, previous and next grease pencil key frame
+- Added a frame grid and a frame grid panel to update the way storyboard frames are placed in 3D space
 
 # 2.0.19 (2022-05-04)